👽 update to the latest mqt-core version (#622)
## Description

This PR updates mqt-core to the latest development version, which
incorporates https://github.com/munich-quantum-toolkit/core/pull/984.

Note that the `mqt-core` Python package reference is updated to a `git`
reference here.
This will not pull from PyPI, but straight from GitHub and build the
package from source.
Before the next QCEC release, this needs to be updated to a stable and
PyPI-available version of mqt-core.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes or removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

Signed-off-by: burgholzer <burgholzer@me.com>