🐛 Fix ancillary handling in ZX checker (#512)
## Description

Fixes #508 

The regression tests and the corresponding CI failures demonstrate the
underlying issue in the ZX checkers:
- the first C++ test, which previously segfaulted, now returns
`equivalent`, while all other checkers agree that the two circuits are
not equivalent.
- the second C++ test is a full C++ reproducer of the failing Python
test. This test segfaults due to a map out of range error.
- the Python test is adapted from #508 

Note the first two tests fail for separate reasons, which means there
are multiple things going wrong here.

This PR adjusts the conditions under which the ZX checker declares it
can handle the underlying circuits, making it more conservative than
before.
Additionally, the out of range error is fixed.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: burgholzer <burgholzer@me.com>
Signed-off-by: Tom Peham <pehamtom@gmx.at>
Co-authored-by: Tom Peham <pehamtom@gmx.at>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>